WIP: Move tests to shared so they are run on android, too
