+++ library/file/src/commonTest/kotlin/io/matthewnelson/kmp/file/NameUnitTest.kt
@@ -17,24 +17,10 @@ package io.matthewnelson.kmp.file
 
 import io.matthewnelson.kmp.file.internal.IsWindows
 import kotlin.test.Test
-import kotlin.test.assertEquals
 
-class NameUnitTest {
+class NameUnitTest: NameSharedTest() {
+    override val isWindows: Boolean = IsWindows
 
     @Test
-    fun givenFile_whenName_thenIsAsExpected() {
-        assertEquals("lorem_ipsum", FILE_LOREM_IPSUM.name)
-        assertEquals("abc123", "/some/path".toFile().resolve("abc123").name)
-        assertEquals("", "".toFile().name)
-        assertEquals(".", ".".toFile().name)
-        assertEquals("..", "..".toFile().name)
-
-        if (IsWindows) {
-            assertEquals("", "\\\\\\".toFile().name)
-            assertEquals("", "C:".toFile().name)
-        } else {
-            assertEquals("", "//".toFile().name)
-            assertEquals("C:", "C:".toFile().name)
-        }
-    }
+    fun stub() {}
 }
