+++ test-android/src/androidInstrumentedTest/fileCommonTestShared
@@ -1,1 +1,1 @@
-../../../library/file/src/commonTestShared/kotlin
+../../../library/file/src/commonTestShared
