+++ test-android/src/androidInstrumentedTest/fileJvmTestShared
@@ -1,1 +1,1 @@
-../../../library/file/src/jvmTestShared/kotlin
+../../../library/file/src/jvmTestShared
